supply/demand cycletime stats
